requirements.txt
aioquic==0.9.25
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.61.1
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.13.0
kiwisolver==1.4.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
MouseInfo==0.1.3
numpy==2.4.2
openai==2.21.0
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pillow==12.1.1
priority==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.19.1
pylsqpack==0.3.19
PyMsgBox==1.0.9
pymssql==2.3.2
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.3.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytweening==1.2.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
selenium==4.30.0
service-identity==24.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
starlette==0.45.3
tqdm==4.67.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wsproto==1.2.0
